+++ ice-20201109/src/Models/CreateEditingProjectRequest.php
@@ -35,19 +35,11 @@ class CreateEditingProjectRequest extends Model
      * @var string
      */
     public $coverURL;
-
-    /**
-     * @description 前端编辑器工程数据结构
-     *
-     * @var string
-     */
-    public $FEExtend;
     protected $_name = [
         'title'       => 'Title',
         'description' => 'Description',
         'timeline'    => 'Timeline',
         'coverURL'    => 'CoverURL',
-        'FEExtend'    => 'FEExtend',
     ];
 
     public function validate()
@@ -68,9 +60,6 @@ class CreateEditingProjectRequest extends Model
         }
         if (null !== $this->coverURL) {
             $res['CoverURL'] = $this->coverURL;
-        }
-        if (null !== $this->FEExtend) {
-            $res['FEExtend'] = $this->FEExtend;
         }
 
         return $res;
@@ -95,9 +84,6 @@ class CreateEditingProjectRequest extends Model
         }
         if (isset($map['CoverURL'])) {
             $model->coverURL = $map['CoverURL'];
-        }
-        if (isset($map['FEExtend'])) {
-            $model->FEExtend = $map['FEExtend'];
         }
 
         return $model;