+++ ice-20201109/src/Models/DeleteEditingProjectMaterialsResponseBody.php
@@ -6,16 +6,16 @@ namespace AlibabaCloud\SDK\ICE\V20201109\Models;
 
 use AlibabaCloud\Tea\Model;
 
-class GetEditingProjectRequest extends Model
+class DeleteEditingProjectMaterialsResponseBody extends Model
 {
     /**
-     * @description 云剪辑工程ID
+     * @description Id of the request
      *
      * @var string
      */
-    public $projectId;
+    public $requestId;
     protected $_name = [
-        'projectId' => 'ProjectId',
+        'requestId' => 'RequestId',
     ];
 
     public function validate()
@@ -25,8 +25,8 @@ class GetEditingProjectRequest extends Model
     public function toMap()
     {
         $res = [];
-        if (null !== $this->projectId) {
-            $res['ProjectId'] = $this->projectId;
+        if (null !== $this->requestId) {
+            $res['RequestId'] = $this->requestId;
         }
 
         return $res;
@@ -35,13 +35,13 @@ class GetEditingProjectRequest extends Model
     /**
      * @param array $map
      *
-     * @return GetEditingProjectRequest
+     * @return DeleteEditingProjectMaterialsResponseBody
      */
     public static function fromMap($map = [])
     {
         $model = new self();
-        if (isset($map['ProjectId'])) {
-            $model->projectId = $map['ProjectId'];
+        if (isset($map['RequestId'])) {
+            $model->requestId = $map['RequestId'];
         }
 
         return $model;